% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCME.R
\docType{data}
\name{CCME}
\alias{CCME}
\title{CCME Freshwater Aquatic Life Guidelines}
\format{
A data frame with 17 rows and 4 variables:
\describe{
\item{parameter}{water quality parameter}
\item{value}{guideline value}
\item{description}{type of guideline}
\item{dependencies}(Y if guideline is calculated based on site-specific conditions (e.g., hardness or ph), N if there is one standard guideline)

...
}
}
\source{
\url{https://www.ccme.ca/en/resources/pollution_prevention.html}
}
\usage{
CCME
}
\description{
A dataset containing the CCME freshwater aquatic life guidelines for various parameters
}
\keyword{datasets}
