% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamGuideline.R
\name{getParamGuideline}
\alias{getParamGuideline}
\title{CCME Freshwater Aquatic Life guidelines}
\usage{
getParamGuideline(param)
}
\arguments{
\item{param}{The name of the parameter, which must be in the exact format/spelling presented in the parameter list in the README.md file}
}
\value{
Generates guidelines values to be presented in the boxplots
}
\description{
The function retrieves guideline values from the CCME table for each matching parameter in your data frame.
This Function is used by the NRV function to render guidelines on boxplots.
You must download the CCME table from the package's git repository and manually edit the site-specific guidelines.
Once you have reviewed the CCME table and edited for your site, then read it into your R session.
Ensure the CCME table is labeled: CCME.
}
