% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M2M_High.R
\name{M2M_High}
\alias{M2M_High}
\title{M2M_High
Calculate upper threshold of Median±2MAD calculation
Function used in NRV calculation function
Function removes NA values}
\usage{
M2M_High(x)
}
\arguments{
\item{x}{Numeric vector of water quality parameter values (results column that includes results below the detection limit as half the detection limit)}
}
\value{
upper threshold of Median±2MAD calculation
}
\description{
M2M_High
Calculate upper threshold of Median±2MAD calculation
Function used in NRV calculation function
Function removes NA values
}
