% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_val.R
\name{max_val}
\alias{max_val}
\title{Calculate the maximum value}
\usage{
max_val(x)
}
\arguments{
\item{x}{Water quality parameter}
}
\value{
Maximum value of water quality parameter
}
\description{
Calculate the maximum value
}
