% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_val.R
\name{min_val}
\alias{min_val}
\title{Calculate the maximum value}
\usage{
min_val(x)
}
\arguments{
\item{x}{Water quality parameter}
}
\value{
Maximum value of water quality parameter
}
\description{
Calculate the maximum value
}
