% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRV.R
\name{NRV}
\alias{NRV}
\title{Natural Range of Variation}
\usage{
NRV(
  inputDatas,
  parameters = NULL,
  generateplots = TRUE,
  renderGuideline = TRUE
)
}
\arguments{
\item{inputDatas}{Single data frame that contains information for each site, or multiple data frames if site data are separated}

\item{parameters}{Default is NULL, which will run the function on each unique parameter in the data frame. Option to select specific parameters from the data frame.}

\item{generateplots}{Default is TRUE, which will generate box plots for each parameter in the data frame (unless specific parameters are provided for the "parameter" argument). Set to FALSE if you do not want box plots generated}

\item{renderGuideline}{Default set to TRUE, which will render all applicable guidelines for each parameter. See the CCME table to confirm the guidelines are appropriate for your dataset. Set to FALSE if the guideline skews the image of the box plot (e.g., guideline is an order of magnitude larger than your maximum value)}
}
\value{
Summary table of results and boxplots for all specified parameters
}
\description{
This function generates a table with the NRV thresholds and associated summary statistics for each parameter in your data frame.
}
\details{
By default, the function also generates box plots for each parameter in your data frame. Can turn off boxplots with generateplots=FALSE (default =TRUE)
You can also select specific parameters to run the boxplots on. E.g., NRV(list(NRV_WC,NRV_TL),"ph","alkalinity_total")
By default, the function renders guidelines from the CCME table (see README file). Turn off guidelines for box plots with renderGuideline=FALSE
You must upload the CCME.csv file from the package's git repository (use readr, change value column to character data type)
}
