% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRV_box_df.R
\name{NRV_box_df}
\alias{NRV_box_df}
\title{NRV Boxplot Data Frame}
\usage{
NRV_box_df(x)
}
\arguments{
\item{x}{Data frame that the summary stats were generated from}
}
\value{
Data frame that the boxplot function can be run on
}
\description{
Create data frame for boxplot function
Parameters need to be presented as individual columns (using the spread functon) after removing redundant columns
}
