% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRV_stats.R
\name{NRV_stats}
\alias{NRV_stats}
\title{NRV Stats Calculation}
\usage{
NRV_stats(x, data_type = "reference")
}
\arguments{
\item{x}{Data frame with water quality data for each site}

\item{data_type}{Data type is set as reference or impacted by the user in the NRV function. Default is reference.}
}
\value{
calculations for the NRV stats table
}
\description{
Calculations that will populate the NRV stats table
}
