% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRV_summ_df.R
\name{NRV_summ_df}
\alias{NRV_summ_df}
\title{NRV summary data frame}
\usage{
NRV_summ_df(
  site,
  date,
  parameter,
  fraction,
  detectionLimit,
  detectionCondition,
  result
)
}
\arguments{
\item{site}{Column with site name or ID}

\item{date}{Column with sampling date}

\item{parameter}{Column with parameter/variable name (e.g., pH, iron, calcium)}

\item{fraction}{Column with analysis fraction details (e.g., total, dissolved)}

\item{detectionLimit}{Column with the value of the detection limit for the analyses}

\item{detectionCondition}{Column that identifies whether the results is below the detection limit. Space should be empty if result is not below the detection limit, if result is below the detection limit should read: below detection limit}

\item{result}{Column with raw result values}
}
\value{
Data frame that subsequent functions can read to calculate the NRV and summary statistics
}
\description{
Prepare your data frame to be read by the NRV package's functions
Data frame needs to be formatted with specific column names and parameter names
Generates new column RDC (resultDetectionCondition) that identifies whether there result is above or below the detection limit
Generates a new column (ResultCalc) that presents the raw result for values above the detection limit, and values that are half the detection limit if result is below the detection limit
This function allows you to enter your current column names to be renamed to fit with the functions
Must enter each column title with the name of the data frame then column title (e.g., df$columnTitle)
make sure detection limit column is numeric
}
