% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{NRVTimeSeries}
\alias{NRVTimeSeries}
\title{Create Time Series Plots for NRV}
\usage{
NRVTimeSeries(dataDF, thresholdDF, paramName = "ALL", season = "Open-Water")
}
\arguments{
\item{dataDF}{data frame containing the time series data}

\item{thresholdDF}{data frame containing the threshold values for the parameters}

\item{paramName}{name of the parameter to plot; "ALL" indicates that plots should be generated for all parameters}

\item{season}{label for the season to be included in the x-axis title}
}
\description{
This function generates time series plots for the specified parameter across all unique parameters
in the dataset or for a specific parameter. It allows the specification of a season,
and displays threshold lines based on input data frames.
}
