% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_W.R
\name{perform_sw_test}
\alias{perform_sw_test}
\title{Perform Shapiro-Wilk Test}
\usage{
perform_sw_test(data_vector)
}
\arguments{
\item{data_vector}{A numeric vector for which the Shapiro-Wilk test will be performed.}
}
\value{
A list with the Shapiro-Wilk test results.
}
\description{
This function performs the Shapiro-Wilk test on a given data vector.
}
\examples{
perform_sw_test(rnorm(100))
}
