% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIF_calc.R
\name{TIF_calc}
\alias{TIF_calc}
\title{TIF Calculation}
\usage{
TIF_calc(x, na.rm = TRUE)
}
\arguments{
\item{x}{Water quality parameter}

\item{na.rm}{Default handle NA values}
}
\value{
upper and lower thresholds of NRV
}
\description{
Used to calculate the lower and upper natural range of variation (NRV) thresholds for a single parameter
}
