% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIF_High.R
\name{TIF_High}
\alias{TIF_High}
\title{Upper Tukey Inner Fence (TIF) Threshold}
\usage{
TIF_High(x)
}
\arguments{
\item{x}{Numeric vector of water quality parameter values (results column that includes results below the detection limit as half the detection limit)}
}
\value{
Upper threshold value
}
\description{
Calculate upper threshold of TIF calculation
Function removes NA values
}
