% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIF_Low.R
\name{TIF_Low}
\alias{TIF_Low}
\title{Lower Tukey Inner Fence (TIF) Threshold}
\usage{
TIF_Low(x)
}
\arguments{
\item{x}{Numeric vector of water quality parameter values (results column that includes results below the detection limit as half the detection limit)}
}
\value{
Lower threshold value
}
\description{
Calculate lower threshold of TIF calculation
Function removes NA values
The calculation returns the value as 0 if the result is a negative number
}
