% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WC_original.R
\docType{data}
\name{WC_original}
\alias{WC_original}
\title{WC_original Example Dataset}
\format{
A data frame with 822 rows and 7 variables:
\describe{
\item{SiteID}{name of the waterbody samples were collected from}
\item{Date}{Date sample was collected}
\item{Variable}{Water quality parameter}
\item{Fraction}{Analysis fraction (e.g., total, dissolved)}
\item{Value}{Analysis result}
\item{Detection}{identifies if the result was "below the detection limit"}
\item{Detection Limit}{Value of the analysis detection limit}

...
}
}
\source{
\url{https://mackenziedatastream.ca/}
}
\usage{
WC_original
}
\description{
A dataset containing the minimum required data to run the NRV function
}
\keyword{datasets}
