% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamGuideline.R
\name{getParamGuideline}
\alias{getParamGuideline}
\title{Get the CCME guideline for the selected parameter}
\usage{
getParamGuideline(param)
}
\arguments{
\item{param}{The name of the parameter, which must be in the exact format/spelling presented in the parameter list in the README.md file}
}
\value{
Generates guidelines values to be presented in the boxplots
}
\description{
#Get the CCME guideline for the selected parameter
}
