% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M2M.R
\name{M2M}
\alias{M2M}
\title{Calculate Median ±2 Mean Absolute Deviation (MAD)}
\usage{
M2M(x)
}
\arguments{
\item{x}{A water quality parameter}
}
\value{
The lower and higher threshold of the natural range of variation for the parameter
}
\description{
Calculate the upper and lower range of the natural range of variation (NRV) for a given parameter with the Median ±2 mean absolute deviation calculation
}
