% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRV.R
\name{NRV}
\alias{NRV}
\title{Natural Range of Variation}
\usage{
NRV(inputData)
}
\arguments{
\item{inputData}{Data frame that contains sample data to be included in the NRV calculation}
}
\value{
Summary table of results all unique parameters
}
\description{
This function generates a table with the NRV thresholds and associated summary statistics for each parameter in your data frame.
}
\details{
By default, the function also generates box plots for each parameter in your data frame. Can turn off boxplots with generateplots=FALSE (default =TRUE)
You can also select specific parameters to run the boxplots on. E.g., NRV(list(NRV_WC,NRV_TL),"ph","alkalinity_total")
By default, the function renders guidelines from the CCME table (see README file). Turn off guidelines for box plots with renderGuideline=FALSE
You must upload the CCME.csv file from the package's git repository (use readr, change value column to character data type)
}
