% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRV.R
\name{NRV}
\alias{NRV}
\title{Natural Range of Variation}
\usage{
NRV(
  inputDatas,
  parameters = NULL,
  generateplots = TRUE,
  renderGuideline = TRUE
)
}
\arguments{
\item{inputDatas}{Single data frame that contains information for each site, or multiple data frames if site data are separated}

\item{parameters}{Default is NULL, which will run the function on each unique parameter in the data frame. Option to select specific parameters from the data frame.}

\item{generateplots}{Default is TRUE, which will generate box plots for each parameter in the data frame (unless specific parameters are provided for the "parameter" argument). Set to FALSE if you do not want box plots generated}

\item{renderGuideline}{Default set to TRUE, which will render all applicable guidelines for each parameter. See the CCME table to confirm the guidelines are appropriate for your dataset. Set to FALSE if the guideline skews the image of the box plot (e.g., guideline is an order of magnitude larger than your maximum value)}
}
\value{
Summary table of results and boxplots for all specified parameters
}
\description{
must run NRV(list(WCTL))
must upload CCME.csv file (use readr, change value column to character data type)
Notes: edit to have 1 dataFrame but multiple sites (i.e., run analyses for each siteID)
Calculate NRV summary table and generate boxplots
Can turn off boxplots with generateplots=FALSE (default =TRUE)
Can turn off guidelines for boxplots with renderGuideline=FALSE (default =TRUE)
}
