% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRV_stats.R
\name{NRV_stats}
\alias{NRV_stats}
\title{NRV Stats Calculation}
\usage{
NRV_stats(x)
}
\arguments{
\item{x}{Data frame with water quality data for each site}
}
\value{

}
\description{
Calculations that will population the NRV stats table
}
