% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIF_High.R
\name{TIF_High}
\alias{TIF_High}
\title{Upper Tukey Inner Fence (TIF) Threshold}
\usage{
TIF_High(x, na.rm = TRUE)
}
\arguments{
\item{x}{The results column that includes results below the detection limit as half the detection limit}

\item{na.rm}{Default handle NA values}
}
\value{
Upper threshold value
}
\description{
Calculate upper thresholds of TIF calculation
}
