% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIF_HighLog.R
\name{TIF_HighLog}
\alias{TIF_HighLog}
\title{Upper Tukey Inner Fence (TIF) Threshold}
\usage{
TIF_HighLog(x)
}
\arguments{
\item{x}{The results column that includes results below the detection limit as half the detection limit}
}
\value{
Upper threshold value that is back-transformed to its original scale
}
\description{
Calculate upper thresholds of TIF calculation on data transformed using the natural logarithm (i.e., log())
}
