% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIF_Low.R
\name{TIF_Low}
\alias{TIF_Low}
\title{Upper Tukey Inner Fence (TIF) Threshold}
\usage{
TIF_Low(x)
}
\arguments{
\item{x}{The results column that includes results below the detection limit as half the detection limit}
}
\value{
Lower threshold value
}
\description{
Calculate upper thresholds of TIF calculation
}
