% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIF_LowLog.R
\name{TIF_LowLog}
\alias{TIF_LowLog}
\title{Lower Tukey Inner Fence (TIF) Threshold}
\usage{
TIF_LowLog(x)
}
\arguments{
\item{x}{The results column that includes results below the detection limit as half the detection limit}
}
\value{
Lower threshold value that is back-transformed to its original scale
}
\description{
Calculate lower thresholds of TIF calculation on data transformed using the natural logarithm (i.e., log())
}
